+++ relay-server/src/statsd.rs
@@ -730,6 +730,11 @@ pub enum RelayCounters {
     ///  - `invalid`: Data was considered invalid and could not be recovered. The reason indicates
     ///    the validation that failed.
     Outcomes,
+    /// The number of individual outcomes including their quantity.
+    ///
+    /// While [`RelayCounters::Outcomes`] tracks the number of times aggregated outcomes
+    /// have been emitted, this counter tracks the total quantity of individual outcomes.
+    OutcomeQuantity,
     /// Number of project state HTTP requests.
     ///
     /// Relay updates projects in batches. Every update cycle, Relay requests
@@ -926,6 +931,7 @@ impl CounterMetric for RelayCounters {
             RelayCounters::BufferProjectChangedEvent => "buffer.project_changed_event",
             RelayCounters::BufferProjectPending => "buffer.project_pending",
             RelayCounters::Outcomes => "events.outcomes",
+            RelayCounters::OutcomeQuantity => "events.outcome_quantity",
             RelayCounters::ProjectStateRequest => "project_state.request",
             #[cfg(feature = "processing")]
             RelayCounters::ProjectStateRedis => "project_state.redis.requests",