+++ tools/bench-buffer/src/main.rs
@@ -237,7 +237,7 @@ fn mock_envelope(
     // Fill with random bytes to get estimated compression ratio:
     let mut payload = vec![0u8; payload_size];
     let fraction = (payload_size as f64 / compression_ratio) as usize;
-    rand::thread_rng().fill_bytes(&mut payload[..fraction]);
+    rand::rng().fill_bytes(&mut payload[..fraction]);
     envelope.extend(payload);
 
     let mut envelope = Envelope::parse_bytes(Bytes::from(envelope)).unwrap();